requirements.txt
selenium
pillow
requests
tqdm